move some shareable function to a utils function file
